Revert "Revert "fix: ensure delete method in reports that have been reported more than once now diminishes in one the reportQuantity instead of deleting the full record""
This reverts commit e6f0d8096e413955cc462e9c44af9a71639cde4f.

+++ api/src/main/java/com/a3bradesco/api/controllers/CnpjReportController.java
@@ -61,14 +61,24 @@ public class CnpjReportController {
 
     @DeleteMapping("/{cnpj}")
     public ResponseEntity<String> deleteReport(@PathVariable String cnpj){
-        cnpjReportService.deleteById(cnpj);
-        CnpjReport isDeleted = cnpjReportService.findById(cnpj);
-        if(isDeleted == null){
-            return ResponseEntity.ok("Denúncia retirada com sucesso!");
-        } else {
-            return ResponseEntity.badRequest().build();
-        }
-        
+    CnpjReport currentDatabaseReport = cnpjReportService.findById(cnpj);
+
+    if(currentDatabaseReport == null) {
+        return ResponseEntity.notFound().build();
     }
-    
+    if(currentDatabaseReport.getReportQuantity() <= 1){
+        cnpjReportService.deleteById(cnpj);
+        return ResponseEntity.ok("Denúncia retirada com sucesso!");
+    } else {
+        CnpjReport newDatabaseReport = 
+        new CnpjReport(
+            currentDatabaseReport.getCnpj(), 
+            currentDatabaseReport.getReportQuantity() - 1, 
+            currentDatabaseReport.getLastTimeReported());
+
+        cnpjReportService.insert(newDatabaseReport);
+
+        return ResponseEntity.ok("Denúncia retirada com sucesso!");
+        }
+    }
 }
